Fix conversion warning on Xcode 5.1

+++ Sources/FKMath.h
@@ -51,7 +51,7 @@ NS_INLINE NSTimeInterval FKTimeIntervalDays(NSTimeInterval days) {
 ////////////////////////////////////////////////////////////////////////
 
 NS_INLINE CGAffineTransform FKAffineTransformMakeRotateTranslate(CGFloat angle, CGFloat dx, CGFloat dy) {
-  return CGAffineTransformMake(cosf(angle), sinf(angle), -sinf(angle), cosf(angle), dx, dy);
+  return CGAffineTransformMake(cos(angle), sin(angle), -sin(angle), cos(angle), dx, dy);
 }
 
 NS_INLINE CGAffineTransform FKAffineTransformMakeScaleTranslate(CGFloat sx, CGFloat sy, CGFloat dx, CGFloat dy) {
